requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto3==1.4.7
botocore==1.7.40
celery==3.1.18
certifi==2018.10.15
chardet==3.0.4
coverage==4.5.2
coveralls==1.5.1
defusedxml==0.5.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==2.1.3
django-allauth==0.37.1
django-cors-headers==2.2.0
django-dotenv==1.4.2
django-rest-auth==0.9.3
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.14
gunicorn==19.7.1
idna==2.6
jmespath==0.9.3
kombu==3.0.37
oauthlib==2.1.0
paystackapi==1.2.7
pluggy==0.8.0
psycopg2==2.7.3.2
py==1.7.0
PyJWT==1.6.4
PyYAML==3.12
python-dateutil==2.7.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.7
redis==2.10.3
requests==2.20.1
requests-oauthlib==1.0.0
s3transfer==0.1.13
six==1.11.0
static3==0.7.0
stripe==2.16.0
tox==2.9.1
urllib3==1.24.1
virtualenv==16.1.0
websocket-client==0.48.0
whitenoise==3.3.1
